requirements.txt
psycopg
psycopg2
pandas
apache-airflow
datetime
requests
pandasql
urllib3
traceback2
apache-airflow
python-dotenv